To remove FortiClient silently
Lots of enterprises may want to remove the FortiClient via inTune or Active Directory, but it is not easy to do it as the FortiClient configuration was protected by the password. You need to unlock the setting before uninstall it. This script is able to unlock the configuration and remove the FortiClient.